fix XPrivToKeyPair to return a keypair instead of a privatekey
